Move subgraph with substitution rule to SubgraphVersionConstraintsFeatureInteractionIntegrationTest
